Base Fragment Implemented
